Refactor modifier package to separate project
