+++ modifier/src/main/scala/co/topl/modifier/BoxReader.scala
@@ -1,6 +1,6 @@
 package co.topl.modifier
 
-import co.topl.modifier.box._
+import co.topl.modifier.box.{Box, BoxId, ProgramBox, TokenBox, TokenValueHolder}
 
 import scala.reflect.ClassTag
