+++ modifier/src/main/scala/co/topl/modifier/box/ArbitBox.scala
