+++ modifier/src/main/scala/co/topl/modifier/box/GenericBox.scala
