+++ modifier/src/main/scala/co/topl/modifier/box/TokenValueHolder.scala
