+++ modifier/src/main/scala/co/topl/modifier/box/serialization/ArbitBoxSerializer.scala
