+++ modifier/src/main/scala/co/topl/modifier/box/serialization/AssetBoxSerializer.scala
