+++ modifier/src/main/scala/co/topl/modifier/box/serialization/BoxSerializer.scala
