+++ modifier/src/main/scala/co/topl/modifier/box/serialization/StateBoxSerializer.scala
